+++ src/auth/AuthButton.tsx
@@ -1,0 +1,11 @@
+export const authButtonstyles = {
+  display: "flex",
+  flex: "1 0 auto",
+  width: "280px",
+  height: "51px",
+  backgroundSize: `280px 51px`,
+  backgroundRepeat: "no-repeat",
+  backgroundPosition: "0",
+  cursor: "pointer",
+  border: "none"
+};