+++ src/auth/BasicAuthButton.tsx
@@ -1,24 +1,26 @@
 /** @jsx jsx */
 import { jsx } from "theme-ui";
 import Button from "components/Button";
-import { modalButtonStyles } from "components/Modal";
 import { OPDS1 } from "interfaces";
+import { authButtonstyles } from "./AuthButton";
 
 const BasicAuthButton: React.FC<{
+  className?: string;
   method: OPDS1.BasicAuthMethod;
   onClick: (type: string) => void;
-}> = ({ method, onClick }) => {
+}> = ({ className, method, onClick }) => {
   const { description, links } = method;
   const imageUrl = links?.find(link => link.rel === "logo")?.href;
   const name = description ?? "Basic Auth";
 
   return (
     <Button
+      className={className}
       aria-label={`Login with ${name}`}
       type="submit"
       onClick={() => onClick(method.type)}
       sx={{
-        ...modalButtonStyles,
+        ...authButtonstyles,
         backgroundImage: `url(${imageUrl})`
       }}
     >