+++ src/auth/SamlAuthButton.tsx
@@ -2,14 +2,17 @@
 import { jsx } from "theme-ui";
 import * as React from "react";
 import Button from "components/Button";
-import { modalButtonStyles } from "../components/Modal";
 import { ClientSamlMethod } from "interfaces";
+import { authButtonstyles } from "./AuthButton";
 
 /**
  * The SAML Auth button sends you off to an external website to complete
  * auth.
  */
-const SamlAuthButton: React.FC<{ method: ClientSamlMethod }> = ({ method }) => {
+const SamlAuthButton: React.FC<{
+  className?: string;
+  method: ClientSamlMethod;
+}> = ({ className, method }) => {
   const handleClick = async () => {
     // get the current location to be redirected back to
     const referrer = encodeURIComponent(window.location.href);
@@ -17,7 +20,11 @@ const SamlAuthButton: React.FC<{ method: ClientSamlMethod }> = ({ method }) => {
     window.open(urlWithReferrer, "_self");
   };
   return (
-    <Button sx={{ ...modalButtonStyles }} onClick={handleClick}>
+    <Button
+      className={className}
+      sx={{ ...authButtonstyles }}
+      onClick={handleClick}
+    >
       Login with {method.description ?? "Unknown IDP"}
     </Button>
   );