+++ src/components/LayoutPage.tsx
@@ -5,9 +5,7 @@ import { AppProps } from "dataflow/withAppProps";
 
 /* LayoutPage is a Page with Header and Footer from Layout, this should be used to wrap pages within the app with sitewide navigation. */
 
-const LayoutPage: React.FC<AppProps> = props => {
-  const { children, library, error } = props;
-
+const LayoutPage: React.FC<AppProps> = ({ children, library, error }) => {
   return (
     <Page library={library} error={error}>
       <Layout>{children}</Layout>